+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostBotLikesPlugin/listeners/ForceAttachLike.kt
@@ -9,15 +9,15 @@ import com.github.insanusmokrassar.TelegramBotAPI.requests.send.SendMessage
 import com.github.insanusmokrassar.TelegramBotAPI.types.ChatId
 import com.github.insanusmokrassar.TelegramBotAPI.types.ParseMode.MarkdownParseMode
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.ForwardedFromChannelMessage
-import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.CommonMessage
-import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.FromUserMessage
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.*
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.TextContent
 import com.github.insanusmokrassar.TelegramBotAPI.utils.extensions.executeUnsafe
-import org.joda.time.DateTime
 import java.lang.ref.WeakReference
 
-private val commandRegex: Regex = Regex("^/attachLikes$")
-private const val commandTemplate: String = "/attachLikes"
+private val attachRegex: Regex = Regex("^/attachLikes$")
+private val attachSeparatedRegex: Regex = Regex("^/attachSeparatedLikes$")
+private const val attachTemplate: String = "/attachLikes"
+private const val attachSeparatedTemplate: String = "/attachSeparatedLikes"
 
 internal fun enableDetectLikesAttachmentMessages(
     adminsHolder: AdminsHolder,
@@ -30,36 +30,52 @@ internal fun enableDetectLikesAttachmentMessages(
         val message = it.data as? CommonMessage<*> ?: return@subscribe
         val userId = (message as? FromUserMessage) ?.user ?.id ?: return@subscribe
 
-        val forwarded = message.forwarded
-        when (forwarded) {
-            is ForwardedFromChannelMessage -> {
-                val originalMessageId = forwarded.messageId
-                if (forwarded.channelChat.id == targetChatId && adminsHolder.contains(userId) && !likesPluginMessagesTable.contains(originalMessageId)) {
-                    botWR.get() ?.executeUnsafe(
-                        SendMessage(
-                            message.chat.id,
-                            "Ok, reply this message and send me `$commandTemplate`",
-                            MarkdownParseMode
-                        )
-                    )
-                }
-            }
-            else -> {
-                (message.content as? TextContent) ?.also {
-                    val reply = message.replyTo
-                    if (commandRegex.matches(it.text) && reply != null && adminsHolder.contains(userId)) {
+        (message.forwarded as? ForwardedFromChannelMessage) ?.let { forwarded ->
 
-                        if (reply.chat.id == targetChatId && reply.messageId !in likesPluginMessagesTable) {
-                            likesGroupsRegistrator.registerNewLikesGroup(
-                                listOf(reply)
+            val originalMessageId = forwarded.messageId
+            if (forwarded.channelChat.id == targetChatId && adminsHolder.contains(userId) && !likesPluginMessagesTable.contains(originalMessageId)) {
+                botWR.get() ?.executeUnsafe(
+                    SendMessage(
+                        message.chat.id,
+                        "Ok, reply this message and send me `$attachTemplate` OR `$attachSeparatedTemplate` to attach with separated message",
+                        MarkdownParseMode,
+                        replyToMessageId = message.messageId
+                    )
+                )
+            }
+
+        } ?: (message.content as? TextContent) ?.let {
+
+            val reply = message.replyTo as? CommonMessage<*> ?: return@let
+            if (!adminsHolder.contains(userId)) {
+                return@let
+            }
+            val text = it.text
+            val realForwarded = reply.forwarded as? ForwardedFromChannelMessage ?: return@let
+
+            if (realForwarded.channelChat.id == targetChatId && realForwarded.messageId !in likesPluginMessagesTable) {
+                when {
+                    attachRegex.matches(text) -> {
+                        likesGroupsRegistrator.registerAttachedLike(
+                            realForwarded.messageId
+                        )
+                        botWR.get() ?.executeUnsafe(
+                            SendMessage(
+                                message.chat.id,
+                                "Likes was attached (can be shown with delay)"
                             )
-                            botWR.get() ?.executeUnsafe(
-                                SendMessage(
-                                    message.chat.id,
-                                    "Likes was attached (can be showed with delay)"
-                                )
+                        )
+                    }
+                    attachSeparatedRegex.matches(text) -> {
+                        likesGroupsRegistrator.registerSeparatedLike(
+                            realForwarded.messageId
+                        )
+                        botWR.get() ?.executeUnsafe(
+                            SendMessage(
+                                message.chat.id,
+                                "Likes was attached by separated message (can be shown with delay)"
                             )
-                        }
+                        )
                     }
                 }
             }